Use userstyle to disable cursor for Chris instead of hardware setting
People often get stuck when they switch to the Chris persona
because they don't know how to get to the selection of profiles
to switch to a different profile. And they forget they could use
the touchscreen.

This changes the simulation to the already existing userstyle instead.
That means people can use the cursor outside the browser,
it is only hidden in the browser.