Make profiles more secure
When lots of people access the profiles on Chromebooks or in the Chrome browser
there is a danger of accidentally saving passwords and other personal information.

This documents how to change the settings to prevent prompts for saving passwords
etc which makes it much less likely that people will accidentally share them.