Fix Simone's scrambling letters messing up Google search results (fixes #84)
When using Simone's scrambling letters userscript, Google search results
are quite badly messed up with icons moving and blinking across the screen.
This is due to the script not just scrambling visible words on the page
but also words within `style` and `script` sections in the HTML.
This fixes that by excluding those from the nodes to scramble.